Классы наследуются от Component
